+++ lib/pages/dashboards/dashboards_carousel_page.dart
@@ -4,7 +4,6 @@ import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/dashboards/dashboard_page.dart';
-import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/sort.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/charts/empty_dashboards_widget.dart';
@@ -43,7 +42,8 @@ class _DashboardCarouselPageState extends State<DashboardCarouselPage> {
         final rangeEnd = getRangeEnd();
 
         return Scaffold(
-          backgroundColor: colorConfig().bodyBgColor,
+          // backgroundColor: colorConfig().bodyBgColor,
+          backgroundColor: Colors.white,
           appBar: const TitleAppBar(title: 'Dashboards'),
           body: CarouselSlider(
             items: dashboards