+++ lib/pages/settings/flags_page.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:intersperse/intersperse.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
+import 'package:lotti/pages/settings/settings_card.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/consts.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
@@ -44,15 +46,18 @@ class _FlagsPageState extends State<FlagsPage> {
           return ListView(
             shrinkWrap: true,
             padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 8),
-            children: List.generate(
-              items.length,
-              (int index) {
-                return ConfigFlagCard(
-                  item: items.elementAt(index),
-                  index: index,
-                );
-              },
-            ),
+            children: intersperse(
+              const SettingsDivider(),
+              List.generate(
+                items.length,
+                (int index) {
+                  return ConfigFlagCard(
+                    item: items.elementAt(index),
+                    index: index,
+                  );
+                },
+              ),
+            ).toList(),
           );
         },
       ),
@@ -95,17 +100,17 @@ class ConfigFlagCard extends StatelessWidget {
     return Card(
       // color: colorConfig().headerBgColor,
       color: Colors.white,
-      elevation: 8,
+      elevation: 0,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8),
       ),
       child: SingleChildScrollView(
         child: ListTile(
           contentPadding: const EdgeInsets.only(
-            left: 16,
+            left: 24,
             top: 4,
             bottom: 8,
-            right: 16,
+            right: 24,
           ),
           title: Row(
             mainAxisSize: MainAxisSize.min,